+++ lib/broadside/error.rb
@@ -1,9 +1,4 @@
 module Broadside
+  class Error < StandardError; end
   class MissingVariableError < StandardError; end
-
-  class Error < StandardError
-    def initialize(msg = 'Broadside encountered an error !')
-      super
-    end
-  end
 end